CMakeLists.txt: search for host tools when cross-compiling
The current toolchain configuration assumes too-strongly that the
builder always wants to use grpc and libprotobuf from the project
submodules. In cross-compilation cases (like when building in OE), we
want to use the host-native tools where possible.

When cross-compiling, prefer to find_*() programs and libraries, rather
than sourcing them from the submodules.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>