CMakeLists.txt: refactor toolchain link logic
The current logic to link library dependencies and other toolchain items
depends upon the CMAKE_CROSSCOMPILING variable, and is somewhat spread
out among the CMakeLists file. CMAKE_CROSSCOMPILING is a misleading
variable name, and the existing logic doens't have good support for
generic linux desktop configurations.

Deprecate the CMAKE_CROSSCOMPILING variable, in favor of a
USE_SUBMODULE_LIBS cmake option. Refactor the linking logic to
be a consolidation of all the linking actions from across the file, and
to better support builds in generic linux environments.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>