CMakeLists: fill out target lib deps
As the test targets are defined, they largely rebuild the code contents
of the device_server code, and so depend upon all the same system
libraries. But the test targets do not declare those dependencies, and
so fail to link.

Shove the ni_grpc_device_server library dependencies into a variable, so
that it can be easily passed along to the test targets.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>